Improved add_questions command
